refactor : test files are now in the same folder but have different name that are more descriptive #46
