+++ 1-DECIDE/test/decideTests.java
@@ -1,4 +1,4 @@
-package test.decide_tests;
+package test;
 import src.CMV;
 import src.PUM;
 import src.FUV;
