+++ 1-DECIDE/test/pumTests.java
@@ -1,4 +1,4 @@
-package test.pum_tests;
+package test;
 import src.CMV;
 import src.PUM;
 import src.FUV;
